+++ clients/serverlessapplicationrepository.d.ts
@@ -36,6 +36,14 @@ declare class ServerlessApplicationRepository extends Service {
    */
   createCloudFormationChangeSet(callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.CreateCloudFormationChangeSetResponse) => void): Request<ServerlessApplicationRepository.Types.CreateCloudFormationChangeSetResponse, AWSError>;
   /**
+   * Creates an AWS CloudFormation template.
+   */
+  createCloudFormationTemplate(params: ServerlessApplicationRepository.Types.CreateCloudFormationTemplateRequest, callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.CreateCloudFormationTemplateResponse) => void): Request<ServerlessApplicationRepository.Types.CreateCloudFormationTemplateResponse, AWSError>;
+  /**
+   * Creates an AWS CloudFormation template.
+   */
+  createCloudFormationTemplate(callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.CreateCloudFormationTemplateResponse) => void): Request<ServerlessApplicationRepository.Types.CreateCloudFormationTemplateResponse, AWSError>;
+  /**
    * Deletes the specified application.
    */
   deleteApplication(params: ServerlessApplicationRepository.Types.DeleteApplicationRequest, callback?: (err: AWSError, data: {}) => void): Request<{}, AWSError>;
@@ -60,6 +68,14 @@ declare class ServerlessApplicationRepository extends Service {
    */
   getApplicationPolicy(callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.GetApplicationPolicyResponse) => void): Request<ServerlessApplicationRepository.Types.GetApplicationPolicyResponse, AWSError>;
   /**
+   * Gets the specified AWS CloudFormation template.
+   */
+  getCloudFormationTemplate(params: ServerlessApplicationRepository.Types.GetCloudFormationTemplateRequest, callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.GetCloudFormationTemplateResponse) => void): Request<ServerlessApplicationRepository.Types.GetCloudFormationTemplateResponse, AWSError>;
+  /**
+   * Gets the specified AWS CloudFormation template.
+   */
+  getCloudFormationTemplate(callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.GetCloudFormationTemplateResponse) => void): Request<ServerlessApplicationRepository.Types.GetCloudFormationTemplateResponse, AWSError>;
+  /**
    * Lists versions for the specified application.
    */
   listApplicationVersions(params: ServerlessApplicationRepository.Types.ListApplicationVersionsRequest, callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.ListApplicationVersionsResponse) => void): Request<ServerlessApplicationRepository.Types.ListApplicationVersionsResponse, AWSError>;
@@ -76,15 +92,17 @@ declare class ServerlessApplicationRepository extends Service {
    */
   listApplications(callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.ListApplicationsResponse) => void): Request<ServerlessApplicationRepository.Types.ListApplicationsResponse, AWSError>;
   /**
-   * Sets the permission policy for an application. See
- Application Permissions
- for the list of supported actions that can be used with this operation.
+   * Sets the permission policy for an application. For the list of actions supported for this operation, see
+ Application 
+ Permissions
+ .
    */
   putApplicationPolicy(params: ServerlessApplicationRepository.Types.PutApplicationPolicyRequest, callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.PutApplicationPolicyResponse) => void): Request<ServerlessApplicationRepository.Types.PutApplicationPolicyResponse, AWSError>;
   /**
-   * Sets the permission policy for an application. See
- Application Permissions
- for the list of supported actions that can be used with this operation.
+   * Sets the permission policy for an application. For the list of actions supported for this operation, see
+ Application 
+ Permissions
+ .
    */
   putApplicationPolicy(callback?: (err: AWSError, data: ServerlessApplicationRepository.Types.PutApplicationPolicyResponse) => void): Request<ServerlessApplicationRepository.Types.PutApplicationPolicyResponse, AWSError>;
   /**
@@ -99,7 +117,8 @@ declare class ServerlessApplicationRepository extends Service {
 declare namespace ServerlessApplicationRepository {
   export interface ApplicationPolicyStatement {
     /**
-     * See Application Permissions for the list of supported actions.
+     * For the list of actions supported for this operation, see Application 
+ Permissions.
      */
     Actions: __listOf__string;
     /**
@@ -146,6 +165,7 @@ declare namespace ServerlessApplicationRepository {
      */
     SpdxLicenseId?: __string;
   }
+  export type Capability = "CAPABILITY_IAM"|"CAPABILITY_NAMED_IAM"|"CAPABILITY_AUTO_EXPAND"|"CAPABILITY_RESOURCE_POLICY"|string;
   export interface CreateApplicationRequest {
     /**
      * The name of the author publishing the app.Minimum length=1. Maximum length=127.Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";
@@ -166,11 +186,11 @@ declare namespace ServerlessApplicationRepository {
     Labels?: __listOf__string;
     /**
      * A local text file that contains the license of the app that matches the spdxLicenseID value of your application.
- The file is of the format file://&lt;path>/&lt;filename>.Maximum size 5 MBNote: Only one of licenseBody and licenseUrl can be specified, otherwise an error will result.
+ The file has the format file://&lt;path>/&lt;filename>.Maximum size 5 MBYou can specify only one of licenseBody and licenseUrl; otherwise, an error results.
      */
     LicenseBody?: __string;
     /**
-     * A link to the S3 object that contains the license of the app that matches the spdxLicenseID value of your application.Maximum size 5 MBNote: Only one of licenseBody and licenseUrl can be specified, otherwise an error will result.
+     * A link to the S3 object that contains the license of the app that matches the spdxLicenseID value of your application.Maximum size 5 MBYou can specify only one of licenseBody and licenseUrl; otherwise, an error results.
      */
     LicenseUrl?: __string;
     /**
@@ -179,11 +199,11 @@ declare namespace ServerlessApplicationRepository {
     Name: __string;
     /**
      * A local text readme file in Markdown language that contains a more detailed description of the application and how it works.
- The file is of the format file://&lt;path>/&lt;filename>.Maximum size 5 MBNote: Only one of readmeBody and readmeUrl can be specified, otherwise an error will result.
+ The file has the format file://&lt;path>/&lt;filename>.Maximum size 5 MBYou can specify only one of readmeBody and readmeUrl; otherwise, an error results.
      */
     ReadmeBody?: __string;
     /**
-     * A link to the S3 object in Markdown language that contains a more detailed description of the application and how it works.Maximum size 5 MBNote: Only one of readmeBody and readmeUrl can be specified, otherwise an error will result.
+     * A link to the S3 object in Markdown language that contains a more detailed description of the application and how it works.Maximum size 5 MBYou can specify only one of readmeBody and readmeUrl; otherwise, an error results.
      */
     ReadmeUrl?: __string;
     /**
@@ -202,11 +222,11 @@ declare namespace ServerlessApplicationRepository {
     SpdxLicenseId?: __string;
     /**
      * The local raw packaged AWS SAM template file of your application.
- The file is of the format file://&lt;path>/&lt;filename>.Note: Only one of templateBody and templateUrl can be specified, otherwise an error will result.
+ The file has the format file://&lt;path>/&lt;filename>.You can specify only one of templateBody and templateUrl; otherwise an error results.
      */
     TemplateBody?: __string;
     /**
-     * A link to the S3 object cotaining the packaged AWS SAM template of your application.Note: Only one of templateBody and templateUrl can be specified, otherwise an error will result.
+     * A link to the S3 object containing the packaged AWS SAM template of your application.You can specify only one of templateBody and templateUrl; otherwise an error results.
      */
     TemplateUrl?: __string;
   }
@@ -293,6 +313,34 @@ declare namespace ServerlessApplicationRepository {
      */
     ParameterDefinitions?: __listOfParameterDefinition;
     /**
+     * A list of values that you must specify before you can deploy certain applications.
+ Some applications might include resources that can affect permissions in your AWS
+ account, for example, by creating new AWS Identity and Access Management (IAM) users.
+ For those applications, you must explicitly acknowledge their capabilities by
+ specifying this parameter.The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM,
+ and CAPABILITY_RESOURCE_POLICY.The following resources require you to specify CAPABILITY_IAM or
+ CAPABILITY_NAMED_IAM:
+ AWS::IAM::Group,
+ AWS::IAM::InstanceProfile,
+ AWS::IAM::Policy, and
+ AWS::IAM::Role.
+ If the application contains IAM resources, you can specify either CAPABILITY_IAM
+ or CAPABILITY_NAMED_IAM. If the application contains IAM resources
+ with custom names, you must specify CAPABILITY_NAMED_IAM.The following resources require you to specify CAPABILITY_RESOURCE_POLICY:
+ AWS::ApplicationAutoScaling::ScalingPolicy,
+ AWS::S3::BucketPolicy,
+ AWS::SQS::QueuePolicy, and
+ AWS::SNS:TopicPolicy.If your application template contains any of the above resources, we recommend that you review
+ all permissions associated with the application before deploying. If you don't specify
+ this parameter for an application that requires capabilities, the call will fail.Valid values: CAPABILITY_IAM | CAPABILITY_NAMED_IAM | CAPABILITY_RESOURCE_POLICY
+     */
+    RequiredCapabilities?: __listOfCapability;
+    /**
+     * Whether all of the AWS resources contained in this application are supported in the region
+ in which it is being retrieved.
+     */
+    ResourcesSupported?: __boolean;
+    /**
      * The semantic version of the application:
  https://semver.org/
  
@@ -313,9 +361,80 @@ declare namespace ServerlessApplicationRepository {
      */
     ApplicationId: __string;
     /**
+     * A list of values that you must specify before you can deploy certain applications.
+ Some applications might include resources that can affect permissions in your AWS
+ account, for example, by creating new AWS Identity and Access Management (IAM) users.
+ For those applications, you must explicitly acknowledge their capabilities by
+ specifying this parameter.The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM,
+ and CAPABILITY_RESOURCE_POLICY.The following resources require you to specify CAPABILITY_IAM or
+ CAPABILITY_NAMED_IAM:
+ AWS::IAM::Group,
+ AWS::IAM::InstanceProfile,
+ AWS::IAM::Policy, and
+ AWS::IAM::Role.
+ If the application contains IAM resources, you can specify either CAPABILITY_IAM
+ or CAPABILITY_NAMED_IAM. If the application contains IAM resources
+ with custom names, you must specify CAPABILITY_NAMED_IAM.The following resources require you to specify CAPABILITY_RESOURCE_POLICY:
+ AWS::ApplicationAutoScaling::ScalingPolicy,
+ AWS::S3::BucketPolicy,
+ AWS::SQS::QueuePolicy, and
+ AWS::SNS:TopicPolicy.If your application template contains any of the above resources, we recommend that you review
+ all permissions associated with the application before deploying. If you don't specify
+ this parameter for an application that requires capabilities, the call will fail.Valid values: CAPABILITY_IAM | CAPABILITY_NAMED_IAM | CAPABILITY_RESOURCE_POLICY
+     */
+    Capabilities?: __listOf__string;
+    /**
+     * This property corresponds to the parameter of the same name for the
+ AWS CloudFormation
+ 
+ CreateChangeSet
+ API.
+     */
+    ChangeSetName?: __string;
+    /**
+     * This property corresponds to the parameter of the same name for the
+ AWS CloudFormation
+ 
+ CreateChangeSet
+ API.
+     */
+    ClientToken?: __string;
+    /**
+     * This property corresponds to the parameter of the same name for the
+ AWS CloudFormation
+ 
+ CreateChangeSet
+ API.
+     */
+    Description?: __string;
+    /**
+     * This property corresponds to the parameter of the same name for the
+ AWS CloudFormation
+ 
+ CreateChangeSet
+ API.
+     */
+    NotificationArns?: __listOf__string;
+    /**
      * A list of parameter values for the parameters of the application.
      */
     ParameterOverrides?: __listOfParameterValue;
+    /**
+     * This property corresponds to the parameter of the same name for the
+ AWS CloudFormation
+ 
+ CreateChangeSet
+ API.
+     */
+    ResourceTypes?: __listOf__string;
+    /**
+     * This property corresponds to the parameter of the same name for the
+ AWS CloudFormation
+ 
+ CreateChangeSet
+ API.
+     */
+    RollbackConfiguration?: RollbackConfiguration;
     /**
      * The semantic version of the application:
  https://semver.org/
@@ -323,11 +442,25 @@ declare namespace ServerlessApplicationRepository {
      */
     SemanticVersion?: __string;
     /**
-     * The name or the unique ID of the stack for which you are creating a change set. AWS CloudFormation generates
- the change set by comparing this stack's information with the information that you submit, such as a modified
- template or different parameter input values. Constraints: Minimum length of 1.Pattern: ([a-zA-Z][-a-zA-Z0-9]*)|(arn:\b(aws|aws-us-gov|aws-cn)\b:[-a-zA-Z0-9:/._+]*)
+     * This property corresponds to the parameter of the same name for the
+ AWS CloudFormation
+ 
+ CreateChangeSet
+ API.
      */
     StackName: __string;
+    /**
+     * This property corresponds to the parameter of the same name for the
+ AWS CloudFormation
+ 
+ CreateChangeSet
+ API.
+     */
+    Tags?: __listOfTag;
+    /**
+     * The UUID returned by CreateCloudFormationTemplate.Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}
+     */
+    TemplateId?: __string;
   }
   export interface CreateCloudFormationChangeSetResponse {
     /**
@@ -348,6 +481,52 @@ declare namespace ServerlessApplicationRepository {
      * The unique ID of the stack.
      */
     StackId?: __string;
+  }
+  export interface CreateCloudFormationTemplateRequest {
+    /**
+     * The Amazon Resource Name (ARN) of the application.
+     */
+    ApplicationId: __string;
+    /**
+     * The semantic version of the application:
+ https://semver.org/
+ 
+     */
+    SemanticVersion?: __string;
+  }
+  export interface CreateCloudFormationTemplateResponse {
+    /**
+     * The application Amazon Resource Name (ARN).
+     */
+    ApplicationId?: __string;
+    /**
+     * The date and time this resource was created.
+     */
+    CreationTime?: __string;
+    /**
+     * The date and time this template expires. Templates
+ expire 1 hour after creation.
+     */
+    ExpirationTime?: __string;
+    /**
+     * The semantic version of the application:
+ https://semver.org/
+ 
+     */
+    SemanticVersion?: __string;
+    /**
+     * Status of the template creation workflow.Possible values: PREPARING | ACTIVE | EXPIRED
+     */
+    Status?: Status;
+    /**
+     * The UUID returned by CreateCloudFormationTemplate.Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}
+     */
+    TemplateId?: __string;
+    /**
+     * A link to the template that can be used to deploy the application using
+ AWS CloudFormation.
+     */
+    TemplateUrl?: __string;
   }
   export interface DeleteApplicationRequest {
     /**
@@ -423,6 +602,50 @@ declare namespace ServerlessApplicationRepository {
      * Version information about the application.
      */
     Version?: Version;
+  }
+  export interface GetCloudFormationTemplateRequest {
+    /**
+     * The Amazon Resource Name (ARN) of the application.
+     */
+    ApplicationId: __string;
+    /**
+     * The UUID returned by CreateCloudFormationTemplate.Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}
+     */
+    TemplateId: __string;
+  }
+  export interface GetCloudFormationTemplateResponse {
+    /**
+     * The application Amazon Resource Name (ARN).
+     */
+    ApplicationId?: __string;
+    /**
+     * The date and time this resource was created.
+     */
+    CreationTime?: __string;
+    /**
+     * The date and time this template expires. Templates
+ expire 1 hour after creation.
+     */
+    ExpirationTime?: __string;
+    /**
+     * The semantic version of the application:
+ https://semver.org/
+ 
+     */
+    SemanticVersion?: __string;
+    /**
+     * Status of the template creation workflow.Possible values: PREPARING | ACTIVE | EXPIRED
+     */
+    Status?: Status;
+    /**
+     * The UUID returned by CreateCloudFormationTemplate.Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}
+     */
+    TemplateId?: __string;
+    /**
+     * A link to the template that can be used to deploy the application using
+ AWS CloudFormation.
+     */
+    TemplateUrl?: __string;
   }
   export interface ListApplicationVersionsRequest {
     /**
@@ -567,6 +790,61 @@ declare namespace ServerlessApplicationRepository {
      */
     Statements?: __listOfApplicationPolicyStatement;
   }
+  export interface RollbackConfiguration {
+    /**
+     * This property corresponds to the content of the same name for the
+ AWS CloudFormation
+ 
+ RollbackConfiguration
+ Data Type.
+     */
+    MonitoringTimeInMinutes?: __integer;
+    /**
+     * This property corresponds to the content of the same name for the
+ AWS CloudFormation
+ 
+ RollbackConfiguration
+ Data Type.
+     */
+    RollbackTriggers?: __listOfRollbackTrigger;
+  }
+  export interface RollbackTrigger {
+    /**
+     * This property corresponds to the content of the same name for the
+ AWS CloudFormation
+ 
+ RollbackTrigger
+ Data Type.
+     */
+    Arn: __string;
+    /**
+     * This property corresponds to the content of the same name for the
+ AWS CloudFormation
+ 
+ RollbackTrigger
+ Data Type.
+     */
+    Type: __string;
+  }
+  export type Status = "PREPARING"|"ACTIVE"|"EXPIRED"|string;
+  export interface Tag {
+    /**
+     * This property corresponds to the content of the same name for the
+ AWS CloudFormation
+ 
+ Tag
+ Data Type.
+     */
+    Key: __string;
+    /**
+     * This property corresponds to the content of the same name for the
+ AWS CloudFormation
+ 
+ Tag
+ Data Type.
+     */
+    Value: __string;
+  }
   export interface UpdateApplicationRequest {
     /**
      * The Amazon Resource Name (ARN) of the application.
@@ -659,6 +937,34 @@ declare namespace ServerlessApplicationRepository {
      */
     ParameterDefinitions: __listOfParameterDefinition;
     /**
+     * A list of values that you must specify before you can deploy certain applications.
+ Some applications might include resources that can affect permissions in your AWS
+ account, for example, by creating new AWS Identity and Access Management (IAM) users.
+ For those applications, you must explicitly acknowledge their capabilities by
+ specifying this parameter.The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM,
+ and CAPABILITY_RESOURCE_POLICY.The following resources require you to specify CAPABILITY_IAM or
+ CAPABILITY_NAMED_IAM:
+ AWS::IAM::Group,
+ AWS::IAM::InstanceProfile,
+ AWS::IAM::Policy, and
+ AWS::IAM::Role.
+ If the application contains IAM resources, you can specify either CAPABILITY_IAM
+ or CAPABILITY_NAMED_IAM. If the application contains IAM resources
+ with custom names, you must specify CAPABILITY_NAMED_IAM.The following resources require you to specify CAPABILITY_RESOURCE_POLICY:
+ AWS::ApplicationAutoScaling::ScalingPolicy,
+ AWS::S3::BucketPolicy,
+ AWS::SQS::QueuePolicy, and
+ AWS::SNS:TopicPolicy.If your application template contains any of the above resources, we recommend that you review
+ all permissions associated with the application before deploying. If you don't specify
+ this parameter for an application that requires capabilities, the call will fail.Valid values: CAPABILITY_IAM | CAPABILITY_NAMED_IAM | CAPABILITY_RESOURCE_POLICY
+     */
+    RequiredCapabilities: __listOfCapability;
+    /**
+     * Whether all of the AWS resources contained in this application are supported in the region
+ in which it is being retrieved.
+     */
+    ResourcesSupported: __boolean;
+    /**
      * The semantic version of the application:
  https://semver.org/
  
@@ -697,8 +1003,11 @@ declare namespace ServerlessApplicationRepository {
   export type __integer = number;
   export type __listOfApplicationPolicyStatement = ApplicationPolicyStatement[];
   export type __listOfApplicationSummary = ApplicationSummary[];
+  export type __listOfCapability = Capability[];
   export type __listOfParameterDefinition = ParameterDefinition[];
   export type __listOfParameterValue = ParameterValue[];
+  export type __listOfRollbackTrigger = RollbackTrigger[];
+  export type __listOfTag = Tag[];
   export type __listOfVersionSummary = VersionSummary[];
   export type __listOf__string = __string[];
   export type __string = string;